moved loaders to showcase / updated docs to reflect changes / updated LoaderCards
