+++ src/app/docs/components/loaders/bounce/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { BounceLoader } from '@/ui/Loader'
+import BounceLoader from '@/showcase/components/loaders/BounceLoader'
 
 <Link className="flex flex-row items-center" href="/docs/components/loaders">
   <ChevronLeft size={17} />
