+++ src/app/docs/components/loaders/classic/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { ClassicLoader } from '@/ui/Loader'
+import ClassicLoader from '@/showcase/components/loaders/ClassicLoader'
 
 <Link className="flex flex-row items-center" href="/docs/components/loaders">
   <ChevronLeft size={17} />
