+++ src/app/docs/components/loaders/neon-glow/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { NeonGlowLoader } from '@/ui/Loader'
+import NeonGlowLoader from '@/showcase/components/loaders/NeonGlowLoader'
 
 <Link className="flex flex-row items-center" href="/docs/components/loaders">
   <ChevronLeft size={17} />
