+++ src/app/docs/components/loaders/page.mdx
@@ -1,17 +1,6 @@
-import {
-  BounceLoader,
-  ClassicLoader,
-  OrbitingLoader,
-  StaggeredFadeLoader,
-  MorphingShapesLoader,
-  NeonGlowLoader,
-  PulsatingGradientLoader,
-  SpinningHelixLoader,
-} from '@/ui/Loader'
-import { CodeGroup } from '@/components/Code'
+
 import { RequiresTailwind, RequiresShadcn } from '@/components/Requires'
-import PreviewCodeTabs from '@/components/CodePreview/page'
-import LoaderCards from '@/components/LoaderCards'
+import LoaderCards from '@/components/cards/LoaderCards'
 
 import { MDXCodeBlock } from '@radix-ui/react-slot'
