+++ src/app/docs/components/loaders/pulsating-gradient/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { PulsatingGradientLoader } from '@/ui/Loader'
+import PulsatingGradientLoader from '@/showcase/components/loaders/PulsatingGradientLoader'
 
 <Link className="flex flex-row items-center" href="/docs/components/loaders">
   <ChevronLeft size={17} />
