+++ src/app/docs/components/loaders/ripple/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { RippleLoader } from '@/ui/Loader'
+import RippleLoader from '@/showcase/components/loaders/RippleLoader'
 
 <Link className="flex flex-row items-center" href="/docs/components/loaders">
   <ChevronLeft size={17} />
