+++ src/app/docs/components/loaders/rotating-dots/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { RotatingDotsLoader } from '@/ui/Loader'
+import RotatingDotsLoader from '@/showcase/components/loaders/RotatingDotsLoader'
 
 <Link className="flex flex-row items-center" href="/docs/components/loaders">
   <ChevronLeft size={17} />
