+++ src/app/docs/components/loaders/spinning-cube/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { SpinningCubeLoader } from '@/ui/Loader'
+import SpinningCubeLoader from '@/showcase/components/loaders/SpinningCubeLoader'
 
 <Link className="flex flex-row items-center" href="/docs/components/loaders">
   <ChevronLeft size={17} />
