+++ src/app/docs/components/loaders/staggered-fade/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
-import { StaggeredFadeLoader } from '@/ui/Loader'
+import StaggeredFadeLoader from '@/showcase/components/loaders/StaggeredFadeLoader'
 
 <Link className="flex flex-row items-center" href="/docs/components/loaders">
   <ChevronLeft size={17} />
