+++ src/components/ComponentCards.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import Link from 'next/link'
 import React from 'react'
-import { OrbitingLoader } from '@/ui/Loader'
+import OrbitingLoader from '@/showcase/components/loaders/OrbitingLoader'
 import { TextTicker } from './PreviewCode/Text'
 import ThreeDButton from '@/showcase/components/button/ThreeDButton'
 import SimpleToggle from '@/showcase/components/toggle/SimpleToggle'
