+++ src/components/cards/LoaderCards.tsx
@@ -1,87 +1,85 @@
 import Link from 'next/link'
 
-import {
-  BounceLoader,
-  ClassicLoader,
-  OrbitingLoader,
-  RippleLoader,
-  RotatingDotsLoader,
-  SpinningCubeLoader,
-  StaggeredFadeLoader,
-  PulsatingDots,
-} from '@/ui/Loader'
+import PulsatingDots from '@/showcase/components/loaders/PulsatingDots'
+import RippleLoader from '@/showcase/components/loaders/RippleLoader'
+import SpinningCubeLoader from '@/showcase/components/loaders/SpinningCubeLoader'
+import RotatingDotsLoader from '@/showcase/components/loaders/RotatingDotsLoader'
+import StaggeredFadeLoader from '@/showcase/components/loaders/StaggeredFadeLoader'
+import OrbitingLoader from '@/showcase/components/loaders/OrbitingLoader'
+import ClassicLoader from '@/showcase/components/loaders/ClassicLoader'
+import BounceLoader from '@/showcase/components/loaders/BounceLoader'
+import NeonGlowLoader from '@/showcase/components/loaders/NeonGlowLoader'
+import PulsatingGradientLoader from '@/showcase/components/loaders/PulsatingGradientLoader'
 
-const data = [
+interface LoaderCardProps {
+  title: string
+  link: string
+  component: React.JSX.Element
+  hide?: boolean
+}
+
+const data: LoaderCardProps[] = [
   {
-    id: 1,
     title: 'Classic',
     link: '/docs/components/loaders/classic',
     component: <ClassicLoader />,
   },
   {
-    id: 2,
     title: 'Bounce',
     link: '/docs/components/loaders/bounce',
     component: <BounceLoader />,
   },
   {
-    id: 3,
     title: 'Orbiting',
     link: '/docs/components/loaders/orbiting',
     component: <OrbitingLoader />,
   },
   {
-    id: 4,
     title: 'Staggered Fade',
     link: '/docs/components/loaders/staggered-fade',
     component: <StaggeredFadeLoader />,
   },
   {
-    id: 5,
     title: 'Rotating Dots',
     link: '/docs/components/loaders/rotating-dots',
     component: <RotatingDotsLoader />,
   },
-
   {
-    id: 6,
     title: 'Spinning Cube',
     link: '/docs/components/loaders/spinning-cube',
     component: <SpinningCubeLoader />,
   },
   {
-    id: 8,
     title: 'Pulsating Dots',
     link: '/docs/components/loaders/pulsating-dots',
     component: <PulsatingDots />,
   },
   {
-    id: 7,
     title: 'Ripple',
     link: '/docs/components/loaders/ripple',
     component: <RippleLoader />,
   },
-
-  // {
-  //   id: 4,
-  //   title: 'Neon Glow',
-  //   link: '/docs/components/loaders/neon-glow',
-  //   component: <NeonGlowLoader />,
-  // },
-  // {
-  //   id: 5,
-  //   title: 'Pulsating Gradient',
-  //   link: '/docs/components/loaders/pulsating-gradient',
-  //   component: <PulsatingGradientLoader />,
-  // },
+  {
+    title: 'Neon Glow',
+    link: '/docs/components/loaders/neon-glow',
+    component: <NeonGlowLoader />,
+    hide: true,
+  },
+  {
+    title: 'Pulsating Gradient',
+    link: '/docs/components/loaders/pulsating-gradient',
+    component: <PulsatingGradientLoader />,
+    hide: true,
+  },
 ]
 
-const HoverAnimationCards = () => {
+export default function LoaderCards() {
   return (
-    <div>
-      <div className="grid grid-cols-2 gap-4 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
-        {data.map((item) => (
-          <Link href={item.link} key={item.id}>
+    <div className="grid grid-cols-2 gap-4 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
+      {data
+        .filter((item) => !item.hide)
+        .map((item, i) => (
+          <Link href={item.link} key={i}>
             <div className="overflow group group rounded-xl border border-white ring-1 ring-zinc-200 transition-all ease-in-out hover:cursor-pointer">
               <div className="flex h-[9rem] items-center justify-center overflow-hidden rounded-t-xl border-b bg-gray-50 text-xs text-gray-400 transition-all ease-in-out group-hover:bg-gray-100 md:h-[12rem]">
                 {item.component}
@@ -92,9 +90,6 @@ const HoverAnimationCards = () => {
             </div>
           </Link>
         ))}
-      </div>
     </div>
   )
 }
-
-export default HoverAnimationCards
