+++ src/showcase/components/loaders/ClassicLoader.tsx
@@ -1,0 +1,5 @@
+export default function ClassicLoader() {
+  return (
+    <div className="h-10 w-10 animate-spin rounded-full border-4 border-gray-200 border-t-red-500" />
+  )
+}
