+++ src/showcase/components/loaders/MorphingShapesLoader.tsx
@@ -1,0 +1,37 @@
+'use client'
+import { motion } from 'framer-motion'
+
+export default function MorphingShapesLoader() {
+  const pathVariants = {
+    hidden: { opacity: 0, pathLength: 0 },
+    visible: { opacity: 1, pathLength: 1 },
+  }
+
+  return (
+    <div className="flex items-center justify-center">
+      <motion.svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="100"
+        height="100"
+        viewBox="0 0 100 100"
+      >
+        <motion.path
+          d="M20,50 Q50,0 80,50 Q50,100 20,50 Z"
+          fill="none"
+          stroke="#FF4136"
+          strokeWidth="8"
+          strokeLinecap="round"
+          variants={pathVariants}
+          initial="hidden"
+          animate="visible"
+          transition={{
+            duration: 1.5,
+            ease: 'easeInOut',
+            repeat: Infinity,
+            repeatType: 'reverse',
+          }}
+        />
+      </motion.svg>
+    </div>
+  )
+}
